@@ -1,8 +1,7 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.1"
+    compileSdkVersion 35
     defaultConfig {
         applicationId "com.example.dark.Surfer"
         minSdkVersion 15
@@ -17,17 +16,19 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+    namespace "com.example.dark.Surfer"
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile 'com.android.support:support-v4:25.0.1'
-    compile 'com.android.support:support-v13:25.0.1'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation 'com.android.support:support-v4:28.0.0'
+    implementation 'com.android.support:support-v13:28.0.0'
     //compile 'com.android.support:appcompat-v7:28.0.0'
-    compile 'com.android.support:support-annotations:25.0.1'
-    compile 'com.android.support:palette-v7:25.0.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    testCompile 'junit:junit:4.12'
-    testCompile 'junit:junit:4.12'
-    testCompile 'junit:junit:4.12'
+    implementation 'com.android.support:support-annotations:28.0.0'
+    implementation 'com.android.support:palette-v7:28.0.0'
+    implementation 'com.android.support.constraint:constraint-layout:2.0.4'
+    testImplementation 'junit:junit:4.13.2'
+    testImplementation 'junit:junit:4.13.2'
+    testImplementation 'junit:junit:4.13.2'
 }
