@@ -7,11 +7,6 @@ import android.content.Intent;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.support.v4.app.NotificationCompat;
-//import android.support.v7.app.NotificationCompat;
-import android.widget.Toast;
-
-import static com.example.dark.Surfer.WebViewActivity.urlAdd;
-import static com.example.dark.Surfer.WebViewActivity.wv;
 
 public class ConnectionBroadReceiver extends BroadcastReceiver {
     String statusOfNetwork;
@@ -23,14 +18,14 @@ public class ConnectionBroadReceiver extends BroadcastReceiver {
             NetworkInfo networkInfo = cm.getActiveNetworkInfo();
             if (networkInfo != null && networkInfo.isConnectedOrConnecting()) {
                 statusOfNetwork = "Internet Connected";
-                if (urlAdd.getText().toString().contains("file:///android_asset/Web/index.html")) {
-                    Toast.makeText(context, "url matched", Toast.LENGTH_SHORT).show();
-                    wv.loadUrl("javascript:history.go(-1)");
-                } else
-                    wv.reload();
+//                if (urlAdd.getText().toString().contains("file:///android_asset/Web/index.html")) {
+//                    Toast.makeText(context, "url matched", Toast.LENGTH_SHORT).show();
+//                    wv.loadUrl("javascript:history.go(-1)");
+//                } else
+//                    wv.reload();
             } else {
                 statusOfNetwork = "Internet Disconnected";
-                WebViewActivity.wv.loadUrl("file:///android_asset/Web/index.html");
+//                WebViewActivity.wv.loadUrl("file:///android_asset/Web/index.html");
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -48,7 +43,7 @@ public class ConnectionBroadReceiver extends BroadcastReceiver {
         //PendingIntent pIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
 
         // Create Notification using NotificationCompat.Builder
-        NotificationCompat.Builder builder = (NotificationCompat.Builder) new NotificationCompat.Builder(context)
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(context)
                 // Set Icon
                 .setSmallIcon(R.mipmap.ic_launcher)
                 // Set Ticker Message
