@@ -1,7 +1,8 @@
 package com.example.dark.Surfer;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
 
 public class Splash extends AppCompatActivity {
 
